[DFT_BLDC_BC_HALL_SAMC21_MCLV2] Primary DFT for BLDC BC HALL SAMC21 project MH3-36512

+++ apps/bldc_bc_hall_sam_c21/firmware/src/config/mclv2_sam_c21_pim/peripheral/adc/plib_adc0.c
@@ -99,7 +99,7 @@ void ADC0_Initialize( void )
     ADC0_REGS->ADC_REFCTRL = ADC_REFCTRL_REFSEL_INTVCC2;
 
     /* Input pin */
-    ADC0_REGS->ADC_INPUTCTRL = ADC_POSINPUT_AIN7;
+    ADC0_REGS->ADC_INPUTCTRL = (uint16_t) ADC_POSINPUT_AIN7;
 
     /* Resolution & Operation Mode */
     ADC0_REGS->ADC_CTRLC = ADC_CTRLC_RESSEL_12BIT | ADC_CTRLC_WINMODE(0) ;
@@ -142,7 +142,7 @@ void ADC0_Disable( void )
 void ADC0_ChannelSelect( ADC_POSINPUT positiveInput, ADC_NEGINPUT negativeInput )
 {
     /* Configure pin scan mode and positive and negative input pins */
-    ADC0_REGS->ADC_INPUTCTRL = positiveInput | negativeInput;
+    ADC0_REGS->ADC_INPUTCTRL = (uint16_t) positiveInput | (uint16_t) negativeInput;
 
     while((ADC0_REGS->ADC_SYNCBUSY & ADC_SYNCBUSY_INPUTCTRL_Msk) == ADC_SYNCBUSY_INPUTCTRL_Msk)
     {
